requirements.txt
Django==1.7
django-crispy-forms==1.4.0
djangorestframework==2.4.3
gnureadline==6.3.3
ipdb==0.8
ipython==2.3.0
wsgiref==0.1.2
